Item Bugs Worked Out

+++ Main Build/Night Defense/NightAttack.cs
@@ -5,7 +5,7 @@ public partial class NightAttack : Resource{
     [Export]
     Godot.Collections.Array<PackedScene> definiteEnemyEncounters;
     [Export]
-    Godot.Collections.Dictionary<double, PackedScene> randomChanceEncounters;
+    Godot.Collections.Dictionary<float, PackedScene> randomChanceEncounters;
 
     //Called from Night Defense on load, meant to check whether something illegal has been placed in one of the data structures
     public void CheckIntegrity(){
@@ -13,7 +13,7 @@ public partial class NightAttack : Resource{
             Node instance = pack.Instantiate();
             if(!(instance is EnemyGroup)) throw new ArgumentException("The Values in Night Attacks's Encounters must be Enemy Groups as Packed Scenes");
         }
-        foreach(KeyValuePair<double, PackedScene> pair in randomChanceEncounters){
+        foreach(KeyValuePair<float, PackedScene> pair in randomChanceEncounters){
             Node instance = pair.Value.Instantiate();
             if(pair.Value is PackedScene && instance is EnemyGroup){
                 instance.Free();
@@ -29,7 +29,7 @@ public partial class NightAttack : Resource{
     public Godot.Collections.Array<PackedScene> GetEnemyGroups(){
         Godot.Collections.Array<PackedScene> finalEnemyGroups = definiteEnemyEncounters.Duplicate();
         Random rand = new();
-        foreach(KeyValuePair<double, PackedScene> pair in randomChanceEncounters){
+        foreach(KeyValuePair<float, PackedScene> pair in randomChanceEncounters){
             if(rand.NextDouble() >= pair.Key) finalEnemyGroups.Add(pair.Value);
         }
         return finalEnemyGroups;
